pyproject.toml
[project]
name = "art-1"
version = "0.1.0"
description = "Yellowstone gRPC Geyser tutorial"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "base58>=2.1.1",
    "grpcio>=1.73.1",
    "grpcio-tools>=1.73.1",
    "python-dotenv>=1.1.1",
    "solders>=0.26.0",
]
